+++ example-app/public/demo-nobundler/layouts/public.js
@@ -1,6 +1,6 @@
 const template = /*html*/`
 <div>
-  <h1>Public Layout</h1>
+  <!-- h1>Public Layout</h1 -->
   <router-view></router-view>
 </div>
 `