+++ example-app/public/demo-nobundler/views/dashboard.js
@@ -1,6 +1,7 @@
 const template = /*html*/`
 <div>
   <h1>Dashboard</h1>
+  <h1 v-for="n of 20">Test</h1>
 </div>
 `